test(activerecord): merge mysql/schema-statements.test.ts into its .trails sibling
